chore: Move types for flags (#4332)
* move types

* fix imports
