+++ app/(gcforms)/[locale]/layout.tsx
@@ -1,7 +1,8 @@
 import { authCheckAndThrow } from "@lib/actions";
 import { ClientContexts } from "@clientComponents/globals/ClientContexts";
 import { ReactHydrationCheck } from "@clientComponents/globals";
-import { FeatureFlags, getSomeFlags } from "@lib/cache/flags";
+import { getSomeFlags } from "@lib/cache/flags";
+import { FeatureFlags } from "@lib/cache/types";
 
 export default async function Layout({ children }: { children: React.ReactNode }) {
   const { session } = await authCheckAndThrow().catch(() => ({ session: null }));
